+++ chaosext/probes.py
@@ -2,3 +2,7 @@
 
 from chaoslib.exceptions import FailedActivity
 from chaoslib.types import Configuration, Secrets
+
+
+def empty_probe(configuration: Configuration = None, secrets: Secrets = None):
+    return True