+++ tests/test_actions.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from chaosext.actions import empty_action
+
+
+def test_empty_action():
+    assert empty_action() is None