+++ tests/test_probes.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+from chaosext.probes import empty_probe
+
+
+def test_empty_probe():
+    assert empty_probe() is True
+