+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 [Unreleased]: https://github.com/chaostoolkit/chaostoolkit-extension-template/compare/0.1.0...HEAD
 
+## [0.1.0][]
+
+[0.1.0]: https://github.com/chaostoolkit/chaostoolkit-extension-template/tree/0.1.0
+
 ### Added
 
 -   Initial release
+-   Added Travis CI
+-   Added empty example tests for probes and actions